requirements.txt
# Web Framework
fastapi
uvicorn[standard]
python-multipart
starlette

# Database
sqlalchemy
aiosqlite
greenlet

# Environment and Configuration
python-dotenv

# HTTP Client & Async
requests
httpx
aiohttp
requests-toolbelt
httpcore
h11
anyio
aiohappyeyeballs
aiosignal
frozenlist
multidict
yarl
idna
certifi
charset-normalizer
urllib3
sniffio
websockets

# PDF Generation
reportlab
jinja2
PyPDF2

# NEO Blockchain (use compatible versions)
neo-mamba
neo3-boa

# Testing
pytest
pytest-asyncio
pytest-mock
# Utilities
pydantic
pydantic-core
pydantic-settings
annotated-types
typing-extensions
attrs
orjson
regex
tenacity
termcolor
nest_asyncio

# LLM Providers
openai
anthropic
tiktoken
tqdm
jiter
distro

# Voice/Speech
elevenlabs

# Google AI
google
google-genai
google-generativeai
google-api-python-client
google-api-core
googleapis-common-protos
protobuf
grpcio

# Vector Database & RAG
pinecone

# Blockchain & Crypto
fastmcp
web3
cryptography
eth-utils
rlp

# Payments
x402

# AWS
boto3
botocore

# CLI Tools
prompt_toolkit

# Other
jsonpointer
propcache
python-telegram-bot
